+++ distribution-lifecycle/src/test/java/io/telicent/smart/cache/distribution/lifecycle/config/TestDistributionLifecycleConfiguration.java
@@ -18,7 +18,10 @@ package io.telicent.smart.cache.distribution.lifecycle.config;
 import io.telicent.smart.cache.configuration.Configurator;
 import io.telicent.smart.cache.configuration.sources.NullSource;
 import io.telicent.smart.cache.configuration.sources.PropertiesSource;
+import io.telicent.smart.cache.distribution.lifecycle.events.listeners.AcknowledgingListener;
+import io.telicent.smart.cache.distribution.lifecycle.events.listeners.LoggingListener;
 import io.telicent.smart.cache.distribution.lifecycle.store.DistributionLifecycleStateStore;
+import io.telicent.smart.cache.sources.kafka.config.KafkaConfiguration;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.AfterMethod;
@@ -28,6 +31,7 @@ import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.file.Files;
+import java.util.Collections;
 import java.util.Properties;
 
 public class TestDistributionLifecycleConfiguration {
@@ -53,10 +57,13 @@ public class TestDistributionLifecycleConfiguration {
 
         // When and Then
         Assert.assertFalse(DistributionLifecycleConfiguration.isEnabled());
-        Assert.assertNull(DistributionLifecycleConfiguration.createDistributionLifecycleStateStore("test"));
+        Assert.assertNull(DistributionLifecycleConfiguration.createStateStore("test"));
         Assert.assertNull(DistributionLifecycleConfiguration.resolveStateFile());
         Assert.assertEquals(DistributionLifecycleConfiguration.resolveListenerThreads(),
                             DistributionLifecycleConfiguration.DEFAULT_LISTENER_THREADS);
+        Assert.assertNull(
+                DistributionLifecycleConfiguration.createTracker(KafkaConfiguration.builder().build(), "test", null, 1,
+                                                                 Collections.emptyList()));
     }
 
     @Test(expectedExceptions = IllegalStateException.class, expectedExceptionsMessageRegExp = ".*Missing configuration.*")
@@ -70,11 +77,11 @@ public class TestDistributionLifecycleConfiguration {
         Assert.assertTrue(DistributionLifecycleConfiguration.isEnabled());
 
         // And
-        DistributionLifecycleConfiguration.createDistributionLifecycleStateStore("test");
+        DistributionLifecycleConfiguration.createStateStore("test");
     }
 
     @Test
-    public void givenFullConfiguration_whenCheckingDistributionLifecycleConfig_thenEnabled_andStoreCanBeCreated() throws
+    public void givenFullConfiguration_whenCheckingDistributionLifecycleConfig_thenEnabled_andComponentsCanBeCreated() throws
             IOException {
         // Given
         File stateFile = Files.createTempFile("distribution-lifecycle-state", ".json").toFile();
@@ -92,9 +99,15 @@ public class TestDistributionLifecycleConfiguration {
         Assert.assertEquals(DistributionLifecycleConfiguration.resolveListenerThreads(), 4);
 
         // And
-        try (DistributionLifecycleStateStore store = DistributionLifecycleConfiguration.createDistributionLifecycleStateStore(
+        try (DistributionLifecycleStateStore store = DistributionLifecycleConfiguration.createStateStore(
                 "test")) {
             Assert.assertTrue(store.activeEvents().isEmpty());
+
+            try (AcknowledgingListener listener = DistributionLifecycleConfiguration.createAcknowledgingListener(
+                    KafkaConfiguration.builder().bootstrapServers("localhost:9092").outputTopic("tests").build(), "test",
+                    "1.2.3", store, new LoggingListener())) {
+                Assert.assertNotNull(listener);
+            }
         }
     }
 
@@ -119,6 +132,21 @@ public class TestDistributionLifecycleConfiguration {
         Assert.assertEquals(DistributionLifecycleConfiguration.resolveListenerThreads(), 4);
 
         // And
-        DistributionLifecycleConfiguration.createDistributionLifecycleStateStore("test");
+        DistributionLifecycleConfiguration.createStateStore("test");
+    }
+
+    @Test(expectedExceptions = NullPointerException.class, expectedExceptionsMessageRegExp = "Kafka Configuration.*")
+    public void givenNullKafkaConfiguration_whenCreatingListener_thenNPE() {
+        // Given, When and Then
+        DistributionLifecycleConfiguration.createAcknowledgingListener(null, "test", "1.2.3", null, null);
+    }
+
+    @Test(expectedExceptions = IllegalStateException.class, expectedExceptionsMessageRegExp = "Invalid configuration for output.*")
+    public void givenEmptyKafkaConfiguration_whenCreatingListener_thenIllegalState() {
+        // Given
+        KafkaConfiguration kafkaConfiguration = KafkaConfiguration.builder().build();
+
+        // When and Then
+        DistributionLifecycleConfiguration.createAcknowledgingListener(kafkaConfiguration, "test", "1.2.3", null, null);
     }
 }